bitswap-monitoring-client: factor out geolocation module
